ARM: DaVinci: reboot via watchdog fix
Before configuring the watchdog timer, the watchdog timer is not
placed into a known state. In other words, if someone happened to use
this timer as a regular timer in an application, the timer is not reset
before it is configured. Hence, I think that we should make sure the
timer is placed into reset before we start to configure the timer.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>
Signed-off-by: Kevin Hilman <khilman@mvista.com>