ARM: DaVinci: i2c fixes from TI
Patch from TI to fix the following i2c issues:

1) Turn on the ARDY interrupt and flag command complete in the
interrupt routine. Without this I2C commands with a repeated start
fail.

2) Fix $B!F(Bi2c_davinci-xfer$B!G(B routine so that it always returns the
number of messages transferred after a successful
transaction. Currently it only does this if there are 2 or more
messages.

3) Change Davinci own slave address as it conflicted with another
device I was using. According to the source comments this was
arbitrarily applied.

Signed-off-by: Kevin Hilman <khilman@mvista.com>