ARM: DaVinci: video capture support
Straight forward port of the VPFE driver from the 2.6.10 kernel,
including some previous work done by Alex Rier. There are a few things
which need cleanup before going upstream.

1) video-buf layer support for contiguous buffers
2) potential reuse of tvp5150 instead of tvp5146

For (2), if anyone is looking for a concrete way to contribute to the
open-source kernel, they could tackle it.  I haven't looked at the
datasheets in detail, but at first glance, I think we could probably
use the existing tvp5150 driver for the tvp5146 instead of adding a
new driver as the current.  The tvp5150 driver may need some minor
adjustments, but that is preferrable than having an entirely different
driver.

Signed-off-by: Kevin Hilman <khilman@mvista.com>