Fix take a long time on IDE probe. fix bug 4678.
Kernel is hanging for a long time on probing the IDE, even
when there's no hard disk connected. This patch fix a bug,
Now we can pass parameters to kernel to skip drive probing
if none exist.
Summary of ide driver parameters for kernel command line:
"hdx=noprobe"  : drive may be present, but don't probe for it