Fix the inverted field hw bug on vid0 plane
This fixes the bug found on 1.3 hw revision. Basically the bug is found when
the video out is in interlaced mode, vid0 picks up the inverted field order
(even when should be odd and vice versa), to avoid that we use the ping-pong
feature of the vid0 plane to swap the order on every interrupt.